update k8s probers
