+++ config/manager-probe-agent/rbac.yaml
@@ -1,27 +1,27 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: kubeprober-readonly
+  name: kubeprober-worker
   namespace: system
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
 metadata:
-  name: kubeprober-readonly-rolebinding
+  name: kubeprober-worker-rolebinding
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: ClusterRole
-  name: kubeprober-readonly-role
+  name: kubeprober-worker-role
 subjects:
   - kind: ServiceAccount
-    name: kubeprober-readonly
+    name: kubeprober-worker
     namespace: system
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
   creationTimestamp: null
-  name: kubeprober-readonly-role
+  name: kubeprober-worker-role
 rules:
 - apiGroups:
   - "*"
@@ -31,6 +31,31 @@ rules:
   - get
   - list
   - watch
-- apiGroups: [""]
-  resources: ["pods/exec"]
-  verbs: ["create"]
+- apiGroups:
+    - ""
+  resources:
+    - namespaces
+    - pods
+    - services
+  verbs:
+    - "*"
+- apiGroups:
+    - "apps"
+  resources:
+    - deployments
+    - replicasets
+  verbs:
+    - "*"
+- apiGroups:
+    - "extensions"
+  resources:
+    - deployments
+    - replicasets
+  verbs:
+    - "*"
+- apiGroups:
+    - ""
+  resources:
+    - "pods/exec"
+  verbs:
+    - "create"
