+++ probers/k8s/control-plane/deplyment_service_checker/namespace.go
@@ -4,13 +4,11 @@ import (
 	"context"
 	"time"
 
-	k8sErrors "k8s.io/apimachinery/pkg/api/errors"
-
 	"github.com/sirupsen/logrus"
 	log "github.com/sirupsen/logrus"
 	corev1 "k8s.io/api/core/v1"
+	k8sErrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
 	"k8s.io/client-go/kubernetes"
 )
