+++ probers/k8s/system/main.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+kubectl probe shell /checkers/system/system.sh
